Mask all alloc-dependant stuff behind alloc feature
This allows the lib to compile for targets that do not have
std library as for example `thumbv6m-none-eabi`.
