+++ src/fft/domain.rs
@@ -12,6 +12,7 @@
 //! This allows us to perform polynomial operations in O(n)
 //! by performing an O(n log n) FFT over such a domain.
 
+#[cfg(feature = "alloc")]
 use super::Evaluations;
 use crate::error::Error;
 #[cfg(feature = "alloc")]
@@ -87,7 +88,6 @@ impl Serializable<{ u64::SIZE + u32::SIZE + 5 * BlsScalar::SIZE }>
     }
 }
 
-#[cfg(feature = "alloc")]
 impl EvaluationDomain {
     /// Construct a domain that is large enough for evaluations of a polynomial
     /// having `num_coeffs` coefficients.
@@ -129,6 +129,7 @@ impl EvaluationDomain {
         self.size as usize
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute a FFT.
     pub(crate) fn fft(&self, coeffs: &[BlsScalar]) -> Vec<BlsScalar> {
         let mut coeffs = coeffs.to_vec();
@@ -136,12 +137,14 @@ impl EvaluationDomain {
         coeffs
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute a FFT, modifying the vector in place.
     fn fft_in_place(&self, coeffs: &mut Vec<BlsScalar>) {
         coeffs.resize(self.size(), BlsScalar::zero());
         best_fft(coeffs, self.group_gen, self.log_size_of_group)
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute an IFFT.
     pub(crate) fn ifft(&self, evals: &[BlsScalar]) -> Vec<BlsScalar> {
         let mut evals = evals.to_vec();
@@ -149,6 +152,7 @@ impl EvaluationDomain {
         evals
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute an IFFT, modifying the vector in place.
     #[inline]
     pub(crate) fn ifft_in_place(&self, evals: &mut Vec<BlsScalar>) {
@@ -166,6 +170,7 @@ impl EvaluationDomain {
         })
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute a FFT over a coset of the domain.
     pub(crate) fn coset_fft(&self, coeffs: &[BlsScalar]) -> Vec<BlsScalar> {
         let mut coeffs = coeffs.to_vec();
@@ -173,6 +178,7 @@ impl EvaluationDomain {
         coeffs
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute a FFT over a coset of the domain, modifying the input vector
     /// in place.
     fn coset_fft_in_place(&self, coeffs: &mut Vec<BlsScalar>) {
@@ -180,6 +186,7 @@ impl EvaluationDomain {
         self.fft_in_place(coeffs);
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute an IFFT over a coset of the domain.
     pub(crate) fn coset_ifft(&self, evals: &[BlsScalar]) -> Vec<BlsScalar> {
         let mut evals = evals.to_vec();
@@ -187,6 +194,7 @@ impl EvaluationDomain {
         evals
     }
 
+    #[cfg(feature = "alloc")]
     /// Compute an IFFT over a coset of the domain, modifying the input vector
     /// in place.
     fn coset_ifft_in_place(&self, evals: &mut Vec<BlsScalar>) {
@@ -194,6 +202,7 @@ impl EvaluationDomain {
         Self::distribute_powers(evals, self.generator_inv);
     }
 
+    #[cfg(feature = "alloc")]
     #[allow(clippy::needless_range_loop)]
     /// Evaluate all the lagrange polynomials defined by this domain at the
     /// point `tau`.
@@ -250,6 +259,7 @@ impl EvaluationDomain {
         tau.pow(&[self.size, 0, 0, 0]) - BlsScalar::one()
     }
 
+    #[cfg(feature = "alloc")]
     /// Given that the domain size is `D`  
     /// This function computes the `D` evaluation points for
     /// the vanishing polynomial of degree `n` over a coset
