+++ src/proof_system/widget/ecc/scalar_mul/fixed_base/verifierkey.rs
@@ -19,6 +19,7 @@ pub(crate) struct VerifierKey {
     pub(crate) q_fixed_group_add: Commitment,
 }
 
+#[cfg(feature = "alloc")]
 impl VerifierKey {
     pub(crate) fn compute_linearisation_commitment(
         &self,
