+++ src/proof_system/widget/logic/verifierkey.rs
@@ -17,6 +17,7 @@ pub(crate) struct VerifierKey {
     pub(crate) q_logic: Commitment,
 }
 
+#[cfg(feature = "alloc")]
 impl VerifierKey {
     pub(crate) fn compute_linearisation_commitment(
         &self,
