+++ src/util.rs
@@ -11,6 +11,7 @@ use dusk_bls12_381::{
 };
 use rand_core::{CryptoRng, RngCore};
 
+#[cfg(feature = "alloc")]
 /// Returns a vector of BlsScalars of increasing powers of x from x^0 to x^d.
 pub(crate) fn powers_of(
     scalar: &BlsScalar,
@@ -42,6 +43,7 @@ pub(crate) fn random_g2_point<R: RngCore + CryptoRng>(
     G2Affine::generator() * random_scalar(rng)
 }
 
+#[cfg(feature = "alloc")]
 /// This function is only used to generate the SRS.
 /// The intention is just to compute the resulting points
 /// of the operation `a*P, b*P, c*P ... (n-1)*P` into a `Vec`.
@@ -54,6 +56,8 @@ pub(crate) fn slow_multiscalar_mul_single_base(
 
 // while we do not have batch inversion for scalars
 use core::ops::MulAssign;
+
+#[cfg(feature = "alloc")]
 pub fn batch_inversion(v: &mut [BlsScalar]) {
     // Montgomery’s Trick and Fast Implementation of Masked AES
     // Genelle, Prouff and Quisquater
